requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.0; python_version >= '3.7'
django-cors-headers==3.11.0
django-dotenv==1.4.2
django==4.0.2
djangorestframework-simplejwt==5.0.0
djangorestframework==3.13.1; python_version >= '3.6'
pillow==9.0.1
psycopg2-binary==2.9.3
pyjwt==2.3.0; python_version >= '3.6'
pytz==2021.3
sqlparse==0.4.2; python_version >= '3.5'
